Expand the responsibilities of Meta
Meta used to be responsible for generating our describe payload, and not much
else. It existed as a consequence of the isolation from the actor-like model
used by DataSets. It now more fully represents a thread-safe snapshot of
DataSets which can be used as required, throughout the system.

The first such usage, the goal of this feature branch, will be to have the
create endpoint parse fields based on the previously parsed status of dataset
columns.